backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
anthropic
python-dotenv
